+++ NEWS.md
@@ -12,6 +12,7 @@
 * `chembl_query()` can now perform both online queries (`mode = "ws"`, default) and offline retrievals (`mode = "offline"`) from a local ChEMBL database. Offline mode currently supports the following resources: `activity`, `assay`, `atc_class`, `binding_site`, `biotherapeutic`, `cell_line`, `chembl_id_lookup`, `compound_record`, `document`, `drug`,`drug_indication`, `drug_warning`, `go_slim`, `molecule`.
 * Results for online and offline queries are identical for most resources. If there are differences, the offline version throws informative warnings.
 * Added a new function `db_download_chembl()` for downloading ChEMBL for fully offline access.
+* Added a new function `chembl_check_db_version()` which retrieves a pinned default ChEMBL database version from .Renviron or .Rprofile and throws an error if none is set. Used for working with pinned versions of ChEMBL in offline mode.
 
 ## OTHER